refactor: move `GeneratorFiles` into separate file
