+++ packages/cli/src/lib/GeneratorFiles.ts
@@ -1,0 +1,5 @@
+import type { URL } from 'node:url';
+
+export type GeneratorFiles = Array<
+  { file: URL; code: string } | { directory: URL; clean: boolean }
+>;
